% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_delete_resolver_endpoint}
\alias{route53resolver_delete_resolver_endpoint}
\title{Deletes a resolver endpoint}
\usage{
route53resolver_delete_resolver_endpoint(ResolverEndpointId)
}
\arguments{
\item{ResolverEndpointId}{[required] The ID of the resolver endpoint that you want to delete.}
}
\description{
Deletes a resolver endpoint. The effect of deleting a resolver endpoint
depends on whether it's an inbound or an outbound resolver endpoint:
}
\details{
\itemize{
\item \strong{Inbound}: DNS queries from your network or another VPC are no
longer routed to the DNS service for the specified VPC.
\item \strong{Outbound}: DNS queries from a VPC are no longer routed to your
network or to another VPC.
}
}
\section{Request syntax}{
\preformatted{route53resolver$delete_resolver_endpoint(
  ResolverEndpointId = "string"
)
}
}

\keyword{internal}
