% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_disassociate_resolver_rule}
\alias{route53resolver_disassociate_resolver_rule}
\title{Removes the association between a specified resolver rule and a
specified VPC}
\usage{
route53resolver_disassociate_resolver_rule(VPCId, ResolverRuleId)
}
\arguments{
\item{VPCId}{[required] The ID of the VPC that you want to disassociate the resolver rule from.}

\item{ResolverRuleId}{[required] The ID of the resolver rule that you want to disassociate from the
specified VPC.}
}
\description{
Removes the association between a specified resolver rule and a
specified VPC.
}
\details{
If you disassociate a resolver rule from a VPC, Resolver stops
forwarding DNS queries for the domain name that you specified in the
resolver rule.
}
\section{Request syntax}{
\preformatted{route53resolver$disassociate_resolver_rule(
  VPCId = "string",
  ResolverRuleId = "string"
)
}
}

\keyword{internal}
