% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route53resolver_operations.R
\name{route53resolver_update_resolver_rule}
\alias{route53resolver_update_resolver_rule}
\title{Updates settings for a specified resolver rule}
\usage{
route53resolver_update_resolver_rule(ResolverRuleId, Config)
}
\arguments{
\item{ResolverRuleId}{[required] The ID of the resolver rule that you want to update.}

\item{Config}{[required] The new settings for the resolver rule.}
}
\description{
Updates settings for a specified resolver rule. \code{ResolverRuleId} is
required, and all other parameters are optional. If you don't specify a
parameter, it retains its current value.
}
\section{Request syntax}{
\preformatted{route53resolver$update_resolver_rule(
  ResolverRuleId = "string",
  Config = list(
    Name = "string",
    TargetIps = list(
      list(
        Ip = "string",
        Port = 123
      )
    ),
    ResolverEndpointId = "string"
  )
)
}
}

\keyword{internal}
